collar/requirements.txt
openai
jpype1